fix: 리뷰 데이터 조회 로직 수정
fix: 리뷰 데이터 조회 로직 수정

+++ src/main/java/com/team/leaf/shopping/product/review/dto/ReviewResponse.java
@@ -12,7 +12,7 @@ public class ReviewResponse {
 
     private long reviewId;
 
-    private long score;
+    private int score;
 
     private String content;
